Moved extension API to a separate module, renamed extension parent module
