+++ symphony-bdk-bom/build.gradle
@@ -23,6 +23,7 @@ dependencies {
     constraints {
         // Internal modules dependencies (Keep them first)
         api "org.finos.symphony.bdk:symphony-bdk-core:$project.version"
+        api "org.finos.symphony.bdk.ext:symphony-bdk-core-extension-api:$project.version"
         api "org.finos.symphony.bdk:symphony-bdk-http-api:$project.version"
         api "org.finos.symphony.bdk:symphony-bdk-http-jersey2:$project.version"
         api "org.finos.symphony.bdk:symphony-bdk-http-webclient:$project.version"
@@ -32,7 +33,7 @@ dependencies {
         api "org.finos.symphony.bdk:symphony-bdk-template-freemarker:$project.version"
         api "org.finos.symphony.bdk:symphony-bdk-template-handlebars:$project.version"
         // extensions
-        api "org.finos.symphony.bdk.ext:symphony-groups-bdk-extension:$project.version"
+        api "org.finos.symphony.bdk.ext:symphony-groups-ext:$project.version"
 
         // External dependencies
         api 'org.projectlombok:lombok:1.18.22'
