+++ symphony-bdk-core-extension-api/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+    id 'bdk.java-library-conventions'
+    id 'bdk.java-publish-conventions'
+}
+
+description = 'Symphony Java BDK Core - Extension API'
+
+dependencies {
+
+    api 'org.apiguardian:apiguardian-api'
+
+}
