+++ symphony-bdk-core-extension-api/src/main/java/com/symphony/bdk/extension/BdkExtensionServiceProvider.java
@@ -1,13 +1,10 @@
-package com.symphony.bdk.core.extension;
+package com.symphony.bdk.extension;
 
 /**
  * {@code BdkExtensionServiceProvider} defines the API or {@link BdkExtension} that wish to provide additional service
  * to BDK developers.
  *
  * <p>Extensions that implement {@code BeforeAllCallback} must also implement {@link BdkExtension}.
- *
- * <p>Concrete implementation often implement {@link BdkAuthenticationAware}, {@link BdkConfigAware}
- * or {@link BdkApiClientFactoryAware} as well.
  *
  * @param <S>
  */
