+++ symphony-bdk-core/build.gradle
@@ -29,6 +29,7 @@ jacocoTestCoverageVerification {
 }
 
 dependencies {
+    api project(':symphony-bdk-core-extension-api')
     api project(':symphony-bdk-http:symphony-bdk-http-api')
     api project(':symphony-bdk-template:symphony-bdk-template-api')
