+++ symphony-bdk-examples/bdk-core-examples/build.gradle
@@ -17,7 +17,7 @@ dependencies {
     implementation 'org.glassfish.jersey.media:jersey-media-multipart'
 
     // let's try an extension
-    implementation project(':symphony-bdk-extensions:symphony-groups-bdk-extension')
+    implementation project(':symphony-bdk-ext:symphony-groups-ext')
 
     runtimeOnly 'ch.qos.logback:logback-classic'
