+++ symphony-bdk-ext/symphony-groups-ext/src/main/java/com/symphony/bdk/ext/group/SymphonyGroupService.java
