+++ symphony-bdk-ext/symphony-groups-ext/src/main/resources/api.yaml
