+++ symphony-bdk-ext/symphony-groups-ext/src/main/resources/symphony-common-definitions.yaml
